feat: finish setting up server project
This commit finishes up the process of setting up the basic server
for Euclid which is all that is inside the scope of #3.
However, it does *not* actually implement any server-side logic
or database structure. In other words, the code and database
as well as Prisma datamodels are all just placeholders for
future server-side work and issues.