feat: add ESLint support for both client and server
Made some changes to the client code in order to match
the Airbnb style guide as much as possible. However,
I am going to *not* follow Airbnb's style guide for
only having JSX in .jsx files as this seems silly.
Their argument is that it is not standard JavaScript
and thus doesn't belong in .js files. However, I personally
don't see this as a problem (given that we have all sorts
of non-standardized JS in our code thanks to Babel...).