docs: add source reference for -diff
The documentation in our `.gitributes` file for how to
ignore a file in diff was lacking. Specifically, it
referenced a Reddit comment rather than a more reliable
and trustworthy source. Essentially, we want to generally
reference the *source* of knowledge rather than someone
talking about it. Thus, this has been remedied by adding
a link to the official Git description of how `.gitattributes` works
and specifically `-diff`.