chore: remove extra .gitignore from src/client
Upon code review I discovered that create-react-app had made
an extra `.gitignore` file in src/client. This is not acceptable
for this project at this time as we wish to only have a single
`.gitignore` file maintained. There may be a point in the future
where we *need* more than one, at which time we will add them.
However, right now one should suffice.