docs: make basic wireframe for Home and Search pages
This commit adds the Design Document for the wireframes
for both the Home and Search pages. So far all of the
wireframes will be in a file called "Wireframe.*". They
are being created and saved in "Wireframe.epgz" which
is a Pencil file. They are exported into a pdf into
the "Wireframe.pdf" file.