feat: begin setting up component hierarchy
This is essentially starting of work done on the branch. To begin
with we are just going to setup some basic placeholder components.
We know from the wireframe that we will need these. For example,
App (routing purposes eventually), Home, Header, Chart, etc.
We are also adding a module resolution alias, specifically: 'App',
since most of the client code will need to refer to App it seemed
silly to use 'screens/App' always instead of 'App'.