feat: start styling home and add recharts package
Started styling the Home page. The header, body, chart, etc. have
all gotten facelifts and look nice. However, there is still no
actual content ( just colored boxes :) ), but this is going to
be addressed with an upcoming commit. In preparation for this
we added the `recharted` package.

This package essentially acts as a React wrapper on top of D3,
which is exactly what we wanted; the charting power of D3 and the
speed and maintainability of React! Next time we will be using it
to add a basic chart which matches the wireframe.