feat: add working chart with mock data
Changes the chart component to actually display a chart! This includes
using: Recharts.js to draw the chart, Faker.js to generate random numbers,
and Moment.js to convert the month numbers to month names. This is doing
essentially exactly what we want. So now we just need to add some real
mock price data for some product X in and start working on drawing
that data.