feat: draw stock data and style chart
Now drawing real stock data! This is much better than trying to
generate data that is similar to stock data. So we are pulling
real stock data from IEX using iex-api. This allows us to work
with data that will look and feel like what our eventual real
data from Amazon, et al. will. In other words, we are now
pulling real info from an API and using real dates and prices.

This means that the structure of our in-progress application
will be able to remain essentially unchanged, since the
underlying data and asynchronous access will be very similar.
We still need to work on formatting the x-axis as it is not
very well laid out. In addition, the tooltip should display
something that is more human readable than the raw day number
like "126"!