feat: start adding GraphQL API for IEX
Previously IEX data was retrieved on the client, we are now retrieving
it on the server instead to prepare for eventual crawling of Amazon,
et al. for data. This allows us to get the interactions between the
client and server as well as database and API mostly correct before
we begin, which is quite helpful since we have *no* historical data
from Amazon to work with during development.