docs: add JSDoc req. to style-guide, trim whitespaces in VS Code
We are now using JSDoc for documenting the code, it is very helpful
to be able to see the signature of a function described in prose
in this way. Furthermore, since JSDoc is widely used and supported
we can actually see this in a tooltip in VS Code! Add some
whitespace trimming settings to VS Code settings list.