feat: setup GraphQL API for IEX on server
This sets up the GraphQL API for accessing IEX on the server.
We are pulling data from IEX and saving to to the PostgreSQL
DB using Prisma. *Note* we still need to handle cases when
the data already exists as it currently gives us an error
saying that inserting the data would violate uniqueness.

Saving a sample query which gets all of the 'AAPL' stocks in
the DB, as this is a good reference on how the client should
eventually use the API.