feat: pull stock info on client from server
This mostly completes the refactoring which was done to move
the pulling of IEX stock data from the client to the server.
The purpose of this refactoring was to prepare for the eventual
pulling of live Amazon data each day and storing this in our DB
which will then be pulled from the client via the server; instead
of just IEX stock data.