feat: make chart tooltip more descriptive
Improves the chart tooltip by making it human readable. Mostly
this involves using moment to format the unix timestamp instead
of just displaying it raw. We are formatting using "MMMM DD, YYYY"
format for now, but this can be changed in the future if needed.
Also added "$" before the price of the stock on the data point
in question just to make it clear that it represents a dollar
price.