feat: add option to change time window
The UX Wireframe included various time windows above the chart
such as: 6 Months, YTD, 1 Year, etc. These are now implemented
in a limited fashion, however it should be noted that the data
in the DB only has up to 2 years, so the windows will not be
fully operational outside of that. As a result, it might make
sense to make changes to the windows, since we won't have
product pricing information anywhere near 5 years or the larger
windows.