refactor: use props instead of React context API
Do to unfamiliarity with the new React context API and the general
problem that I was trying to solve I used the React Context API
for setting the time window. There is an issue where the user
gets a "Loading" message if they click the same time window button
too many times and I am investigating this.

During this process, I realized that it is overkill and perhaps even
an anti-pattern to use the React context API for passing down the
time window when a simple prop will do. Thus, this commit does
this refactoring/simplification.