fix: resolve issue with multi-clicking breaking time window
Time window selection was broken because of what appears to be
an issue with using React with Apollo Client's Query Component.
Essentially, if it gets re-rendered by React due to state/prop change
and it runs the query it can sometimes come back with a broken result,
meaning loading is true, error is undefined, etc. This seems to be very
odd behavior, but it also is rather consistent.

Due to this consistency we are able to use a workaround where we
just check if the error state is undefined and then only show loading
if this is false. However, there are some edge cases that this
introduces so we also need to setup a dummy object representing
a basic state to handle this case.

Essentially, what seems to be happening is right at the beginning
of setting up the component tree the Query component hasn't been setup,
so it seems to get into the same error state, as before. However, since
this is actually expected if we just give it a dummy object with
pre-filled data then this can be fixed. Since React will immediately
give it the prop and then it will re-render it again and run the
Query which will give it back the real data and then it will show
as correct to the client.