fix: change x-axis tick formatter based on time window
Some of the time windows need different tick formatters since
they can potentially have dramatically different time ranges.
For example, YTD in January is basically < 1 month of data, whereas
3 Years can be ~36 months of data!

The original month formatter works well for what it was designed
for, which was 2 years worth of data, but doesn't work very well
for YTD in January or 6 Months (which has some duplicated months).

This commit resolves these issues by adding days into the mix as
well to make each of the ticks have distinct labels in each of
these cases. In addition, in the very small ranges (like < 1 month
for YTD in January) we also use the full name of the month just
to have the tick label take up more space, to not have lots
of empty space.

This commit also changes the format of the date on the tooltip
slightly by removing the zero-padding. That is, "January 09" becomes
"January 9". This seems reasonable and logical to do as it doesn't
really add anything and actually even slightly hurts readability.
For these reasons, it was done on the tick formatter, so for
consistency we will just use this format/date style everywhere.