feat: add more data to DB and make func to populate DB for ticker
Since the front-end chart component is now essentially complete,
we have a lot more options to work with to view data. So we needed
to expand the range of data we had from 2 years to 5 years. In
addition, since we are now working on making the search field
we also need data to search for (rather than always displaying AAPL)
so we are also adding some data for other tickers to the DB.