fix: make server ip variable instead of hard-coded `localhost`
Due to mostly testing on my local development machine I had
overlooked the fact that I had hard-coded `localhost` for the
Apollo Client server connection. This caused the site to get a
network error on *any* other machine! This is now changed to be
a variable with the local IP address of my development machine.

As a result, it now works on any computer on my LAN, however
it *still* will not work on any computer on a network! We need
a solution for this, preferably involving a docker variable or
something of that nature setup by the deployment system. However,
this will need to wait until we actually start working on ops
tasks and start doing deployment to production systems.