chore: use babel-eslint parser instead of eslint
The standard ESLint parser cannot handle ESNext very well.
For example, it doesn't handle the class property syntax
for defining methods. That is, using arrow functions to
define methods on a class rather than the standard
ES2015 way (which requires a .bind(this) call in the constructor).

This is a Stage 3 ESNext proposal (proposal-class-fields),
and is supported by babel using by using the plugin
@babel/plugin-proposal-class-properties. However, ESLint does
not support it. A Stage 3 proposal at this point can be
considered to be pretty well standardized (though anything
less should *not* be). So we should be able to use it
in the project.

The babel-eslint parser is an alternative which allows use of
ESNext syntax such as this while also still using ESLint.