fix: make some styles on the Chart and ChartGroup more consistent
Make some of the styles on the Chart and ChartGroup more
consistent with the rest of the styling of the project. There
were a few areas that weren't very well styled, due to focusing
more on functionality at the start rather than appearance.
However, now that the functionality is essentially done it seemed
prudent to go ahead and fix this.