refactor: remove StockList, and move DB to docker
Remove StockList as an actual Table in the DB and have it just
be a query instead. This required a huge refactoring of nearly
the entire codebase! Since we were using StockList in the Client
on the Chart component, and in the Server where we pull data and
also saving it, as well as the DB model and GraphQL interface.

As a result, nearly every part of the codebase is affected by
this change. This is a simple change, but it is also a good
indicator of just how much work is required to do something
like change the fundamental structure of the database, even
if no data is removed or even significantly altered.

During the work on this change I had to change the way the
stock items were being added to the DB through Prisma.
Previously they were being handled internally using the `create`
keyword. However, now it is up to the code to handle this as that
was only possible due to having StockList be a table that was easy
to create manually and having Prisma handle the actual stocks.

As a result, we are doing some async work to actually save data
to the DB now. During this process the DB got swamped and as
part of the fix for this I ended up switching from using
PostgreSQL on the local development system to using it from
a docker instance. This is actually quite good as it was required
eventually regardless. The fix for this problem, which was that
we were hitting the DB connection limit, was to limit the amount
of requests being made by waiting for the requests to complete
before spamming more (that is do it sequentially or
semi-sequentially). More specifically, we are doing up
to 50 requests at a time (with Async's eachLimit function),
instead of potentially thousands.

BREAKING CHANGE: StockList no longer exists and is now just
a query which can gotten with a where clause on stocks.