feat: make search field set the chart to a new stock
This a feature that has been in the works for awhile now, the
basic idea is to provide a very simple UI for switching the stocks
that are being displayed on the chart. Until now, it has always
been hard-coded to 'AAPL'. However, with this change you may
now select any ticker from the search box and it will change
the chart.