build: run ESLint and Jest before allowing commits
As part of an ongoing effort to raise the quality standards, all
commits must pass ESLint and Jest/testing before they can be made.
Until now, ESLint was done on a "gentleman's agreement" where
I simply fixed all of the errors that I saw, rather than requiring
it for everything. Meaning, the scripts created by
create-react-app did not pass ESLint!

Furthermore, this issue was problematic because there was some
asymmetry due to the scripts from create-react-app requiring
ESLint before the React code could be built, but no such rule
existing on the serve-side. So this framework will fix this
problem. However, it should be noted that this commit does
*not* actually add ESLint or testing requirements to the
server, as this will be done in an upcoming commit.