build: run lint and tests on push instead of commit
Make the git process less painful by only running tests
and lint on push instead of on commit. This was tending to
penalize making frequent commits which is *not* good!

On the other hand, we do want to make sure that all code that
actually gets pushed to repos is tested and linted so this
is a good alternative.

Also fix a failing PropTypes issue.