chore: run static checking and testing on commit and push
In order to help ensure that no commits break the build, we
will now be running static checking and testing on every
commit as well as on push. I tried to hold off on this at
first, but since we could potentially have a commit that
breaks the build, which is not the latest one, we must.

That is, it is possible for a build to pass the push, but
not all commits, if you have an error and then fix it
without squashing before pushing. Since we are not
doing squashing in the repo this is the only way to help
ensure that no commits break the build, and thus
`git bisect`.