refactor: extract tsconfig options into a base file
As part of the TypeScript standardization efforts begun with
`tslint-config-euclid`, I am moving the tsconfig compiler flags
to their own base file. This will allow us to change the options
in a single place and have this be reflected throughout the
codebase, rather than having to manually keep track of which options
are enabled where and copy-paste them around.

This is also nice because it guarantees that the options are the
same throughout, so that the code is all on the "same" version
of the TS compiler.