build: make TypeScript compiler more strict
Help to ensure that the TypeScript code quality is high by
enabling all of the strict flags, and things like `utf-8`
which will help support emojis in the code. This didn't actually
break any of the files, linting, or tests, which is a good sign
that the code was already of a decently high quality.
However, it is always good to be sure.