build: stop compiling TypeScript on build, remove loading msg
The `create-react-app` scripts were detecting that TypeScript
is in use and were compiling all of the TypeScript files on build.
This is *not* what we want as it is far too slow, instead we are
wanting to have babel throw away the types and do our own periodic
type checking.

Also remove the loading message that was displayed when fetching
data for the chart from the server, this made the chart flash
on and off instead of naturally interpolating between states.