fix: fix broken build, add build script to pre-commit
Remove TypeScript in an earlier commit from the build on the
client, broke it! However, this was not caught by any of the
pre-commit scripts, which is very disturbing. To remedy this
I have added a 'build' section to all pre-commit scripts.

In other words, every pre-commit script first checks that the
code actually builds, before even trying to do type-checking
and linting. Since in this case above, both of the latter
passed, but not the build itself. This should help ensure
that nothing breaks the build!