test: add tests for `pullAndSaveData` and make IEX util
In the first pass of the PR there were no tests for
`pullAndSaveData`, however during code review there was a
question of whether or not there was a bug in this file.
Therefore, tests were added to answer this question in a robust
way. In the process, a bit of refactoring was done to
the file also since there were some style concerns that needed
to be addressed.

A utility for mocking the IEX API site was added as well due to
needing this in two test files.