style: use `no-unsafe-any` TSLint rule on server side
Use `no-unsafe-any` rule on server side to help catch potential
errors. Just during this process, the type-checker found several
errors that I had missed! Thus, this is definitely a good rule
to use. Starting with `no-unsafe-any` instead of both it
and `no-any` because `no-unsafe-any` seems to catch most of the
major errors that we care about, without being a huge burden
to implement.