style: add `no-any` on the server side as a warning
The `no-any` rule is a bit heavy-handed to be used as an error.
However, it is good to have notifications that we are using
any in our code. Thus, enabling it as a warning seems to a good
approach. This way, it won't fail the build, but it will notify
developers that the code could potentially be improved by
replacing these any types with real ones.