feat: enable `no-unsafe-any` as error on client
To help ensure that we are actually using the benefits of type
checking from TypeScript we now have `no-unsafe-any` as an error
on both the client and server. In addition, `no-any` is a warning
to serve as a heads-up that we are bypassing the type system,
however as long as it isn't explicitly unsafe we won't make it
an error for now.