test: add more unit tests on the client
Begin testing the client in earnest. Test: ticker action creator,
TimeWindow, ServerInfo, and the new utility isObjectConstant.
This utility was created to specifically test whether or not an
object is truly constant; is frozen, has a null __proto__, and
all descendents are the same. This is very helpful as it allows
us to essentially assert that an object is *really* constant,
and thus "test" our constants, such as TimeWindow and ServerInfo.