refactor: split up massive Chart component for testing
The Chart component was huge and very tightly coupled, this is
bad for many reasons, the most practical of which is that this
issue deals with testing and the coupling makes testing unit
testing this logic nearly impossible.

Testing components themselves will be handled in a future issue,
however the business logic that was in Chart needs to be tested.
Thus, the refactoring was necessary. So many functions were
split up and created in new files, making the Chart component
short, lean, loosely-coupled, and readable!

These functions are documented thoroughly with JSDoc and will
be tested in a future commit.

Also got the Jest extension working with the Euclid client!
Still need to come up with a solution for making it work with both
the client **and** server, but this is a good first step
regardless.