build: start using docker/Prisma server in VM instead of locally
Moved the Prisma server/PostgreSQL DB to their own VM and switched
from using Docker Desktop to using Docker Toolbox. This allows
me to disable Hyper-V and get back to using VMs in VirtualBox!
A side-benefit of this is the fact that many implicit dependencies
in the project which assumed that the Prisma server was running
on localhost have been removed.

This will allow us to have a much easier time migrating to
product where we will have the Prisma server hosted on some
random endpoint that will almost certainly not be the same
machine as our Node.js server. As part of this I removed the
"GraphQL" item from the ServerInfo-secret files as it was
confusing and not helpful, since there is no difference between
GraphQL and Node in our situation so far, and nothing should
be accessing the Prisma server except `graphql-yoga` through
the generated Prisma client lib.