test: add cypress e2e for SearchField dropdown boxes
This adds Cypress end-to-end tests for the SearchField component
which is just some `react-select` components. This passes locally,
however this commit does *not* include any incorporation into the
testing scripts or CI scripts. This is going to require more more
work due to changes required to the server to send 200 status
codes so that we can detect when it is finished starting and move
on to the testing.