test(client): use Cypress to test React components
Now testing React components using Cypress, this seems to be the
best way to test React as some components such as `react-select`
do not lend themselves to being easily unit tested by tools such
as Jest and Enzyme.

Cypress allows testing of React components as if the user is
actually using it, and can be run headlessly for easy integration
into scripts.

This commit only tests the client's SearcField component,
specifically the dropdown boxes. The chart will need to be tested
in an upcoming sprint, issue, and commit.