feat: write integration test for header component
This is essentially a proof of concept of how to do integration tests
on components which are in Storybook. That is, just taking the e2e tests
for the Header component and running them on the Storybook version. This
*should* have been very simple, basically a copy-paste from the e2e
test, however it was not.

There is an issue with Cypress, namely that [it doesn't properly support
iframes.](https://github.com/cypress-io/cypress/issues/136)
To work around this issue, I investigated many workarounds and came up
with a solution that I think solves the problem in a very elegant way:
load the iframe with Cypress directly. This way, we can just use Cypress
as normal and completely ignore the iframe issue, however this did
require a bit of work figuring out the best way to load the iframe and
selecting it.

There was a bit of a problem with getting the src url of the Storybook
preview iframe which was 'id=+' for some reason instead of the real
id of the div. This may be a bug caused by Storybook 5.0 which was just
released. Either way, I worked-around this issue by selecting the
id from the story div using a Cypress alias and then manually telling
Cypress the url for the iframe. This is not really great as it is
kind of a hacky solution, however it does solve the problem in a way
that is pretty unlikely to break in the future (since I cannot imagine
Storybook changing their id format, as it is pretty consistent with
the UI).

Signed-off-by: David Piper <david.piper@mailfence.com>