ci: fix broken ci which was caused by a node 11.11.0 problem
--------------------------------------------------------------
                squashed commit messages
--------------------------------------------------------------
ci: disable cache to attempt to fix failing ci

CI started failing for no apparent reason, the same scripts pass
locally with no problems and no changes have been made to the
affected areas. Specifically, the jest unit tests on the client
are failing. Even though nothing should have changed there.

So my theory is it might be cache related.

Signed-off-by: David Piper <david.piper@mailfence.com>

ci: fix broken ci which was caused by a node 11.11.0 problem

Travis CI was using the latest version of node, which was recently
updated to 11.11.0, [this version broke Jest](https://github.com/facebook/jest/pull/8050).
So I updated to the latest version of Jest which includes this fix,
however this has made it clear to me that we should be using stable
versions of Node rather than Latest for CI (pre-production) and
production.

Having, the CI just randomly break due to bugs in libraries is
completely unacceptable. Thus, I have also switched the node version
that Travis uses to `lts/*' so that something like this will hopefully
not happen again.

Signed-off-by: David Piper <david.piper@mailfence.com>

--------------------------------------------------------------

Signed-off-by: David Piper <david.piper@mailfence.com>