style: use `tslint-config-prettier` so that prettier doesn't break lint
Prettier is "pretty" good, however it has a very simplistic approach
to code-formatting and is not very configurable. This means that
it breaks tons of linting rules and style-guides. This is unfortunate
as it means that developers have to choose between the style-guide
and prettier.

However, since prettier does do a good job of code-formatting I think
that it is acceptable to use its formatting despite it not following
airbnb (the usual style guide for Euclid).

Signed-off-by: David Piper <david.piper@mailfence.com>