chore: update monday; update all npm packages
------------------------------------------------------
                squashed commits
------------------------------------------------------

chore: update root packages and server packages

Signed-off-by: David Piper <david.piper@mailfence.com>

chore: update client packages with `npm update`

Having a broken build when I upgrade with `ncu -u` so going to have to
do a bit more investigation into which package is breaking things.
For some reason it is saying "webpack is not a function."

Signed-off-by: David Piper <david.piper@mailfence.com>

chore: finish upgrading client, found and fixed breaking package

Finish upgrading the client packages, `react-dev-utils` 8.0 introduced
a breaking change in the signature of `createCompiler` it now takes
an options object instead of multiple arguments. This change was made
due to work on `create-react-app` to [speed-up TypeScript performance](https://github.com/facebook/create-react-app/pull/6406).

It is probably worth investigating these options in the future as well.
Currently, we have disabled TypeScript at build-time/run-time due to
the aforementioned issue, in favor of running type-checking as part of
the linting process. However, it is probably a good idea to see what the
`create-react-app` team did and see if it is worth integrating it into our
build/run process.

Having type-checking as part of the build/run process might help to
catch errors a bit more quickly, rather than catching them in the
editor or as part of pre-commit we could catch them right after
they are made during initial development.

Signed-off-by: David Piper <david.piper@mailfence.com>

------------------------------------------------------

Signed-off-by: David Piper <david.piper@mailfence.com>