build: switch root tasks to use gulp instead of npm scripts
Switched the root tasks from npm scripts to gulp, this was quite
involved as these tasks were very complex and interrelated. However,
they are _much_ simpler now since using gulp. There are a few
gotchas involving my new `terminal-spawn` library, basically
some of the processes don't terminate easily; specifically the server.

This is not an issue in this commit, but it is something to watch
out for when working with this.

Signed-off-by: David Piper <david.piper@mailfence.com>