build: change client to use gulp
This commit finishes up the npm scripts -> gulp conversion process.
However, it should be noted that there is one instance when npm
scripts still need to be used: when running a server that needs
to be killed (such as in the e2e tests). Otherwise, gulp is
used and _should be_ used for all tasks.

Note that the way that commands are run is still done via
npm scripts, that is the "API" of the build process is still
exposed via npm. Thus, for example to run the pre-commit scripts
from the terminal you would run `npm run preCommit`. This just
_happens_ to call a gulp task via `npx gulp preCommit`. This way
the build process is more familiar, and users/me don't have to
think about the fact that the tasks are running in gulp
instead of npm.

This is also nice because oh-my-zsh provides a [npm auto-completion
plugin](https://github.com/robbyrussell/oh-my-zsh/wiki/Plugins#np://github.com/robbyrussell/oh-my-zsh/kiwi/Plugins#npm)
but as far as I can tell does _not_ provide a gulp one. So this
makes using tasks much more user-friendly. Furthermore, this
method means that there is a clear distinction between
"public" tasks and "private" internal ones. The public ones
are exported and have an npm script API. The private ones are
_not_ exported and don't have a npm script API.

Another benefit of this approach, is that the npm script API
tasks can have **any** name desired, rather than the severe
restrictions that come from using functions as task using
the modern gulp style. This means that the gulp task back-end
can have function style names (no colons or dashes in the names)
and the npm script task front-end can have standard script names
(with colons and dashes).

Signed-off-by: David Piper <david.piper@mailfence.com>