build: refactor the build to use dotenv file
Refactor the build to use environment variables stored
in a dotenv file. This allows the things like the
address of the server, prisma, storybook, etc. to be
stored independently of the code. Thus, switching
any of these is now as easy as modifying this file.

This commit is quite large, as this refactoring
affected lots of files in the code-base, since
the information now in env vars was hard-coded
all over the place.

- create dotenv file that stores deployment-specific information
- change code in root, server, and client to use dotenv file
- use dotenv in client code by adding webpack plugin which does
textual-replacement at bundle-time
- change build scripts spawning processes to use this information,
avoid duplication by creating constant files which contain this
information thus keeping the build DRY
- make "template" files which contain Prisma information with
pseudo-env-vars in them
- make scripts to generate "real" files based on the template files
by doing a textual-replacement on them, filling in the env vars
from the dotenv file
- create groups of dotenv files to track the deployment information
- replace .babelrc files with babel.config.js as this is more flexible

There are lots of changes in this commit, I wish that it made
sense to split it up, however these changes really all belong
together as they are all related to the dotenv refactoring.
Still it would be about the same if I did a squash, so this
is fine.

Signed-off-by: David Piper <david.piper@mailfence.com>