build: add slack notifications on verify script finish
Add slack notifications on verify script, which is the
new name of the pre-commit script. This is to match
the behavior of the CI, so that I can know when the
verify script is done so that I can resume working.

Caveat:

There is no notification on an error right
now due to technical limitations in gulp, I have
a potential solution involving preemptively catching
the errors, and sending the notification.

Signed-off-by: David Piper <david.piper@mailfence.com>