refactor: move prisma version to env var, update packages
Move the Prisma version into an environment variable so
that only that needs to be changed, rather than changing
various files all over the place which reference it.

Update npm packages with `npm update` in: root, client,
and server.

Still need to do full update (outside semver ranges) with
ncu.

Signed-off-by: David Piper <david.piper@mailfence.com>