test: load sb url in cypress from dotenv, start visual testing chart
Load the Storybook url in cypress from dotenv file rather than
hard-coding it. Begin testing the chart component with cypress +
visual regression testing. This is still quite preliminary as
it doesn't actually wait for the chart to load before testing,
thus it is just testing the background styling :). However, this
is a good first step towards getting the last thing tested,
the chart!

Signed-off-by: David Piper <david.piper@mailfence.com>