test: test the default chart view story with visual regression
This is the first part of the cypress visual regression testing
implementation, unlike the one from yesterday this one actually
works! I ended up using a different module, as the other one
had only one release and no updates; meaning it is likely
dead and unsupported...

This required a small change to the client to implement, as
Cypress had no way to detect that the chart animation was
done (apparently the HTML doesn't change at all...) so I added
an invisible span which gets a special marker class when the
animation is done. This is possible because Recharts,
fortunately, provides an `onAnimationEnd` callback prop.

Signed-off-by: David Piper <david.piper@mailfence.com>