test: test tooltip, fix race-condition in chart visual tests
The tooltip on the chart had yet to be tested until now. In addition,
during the process of adding tests for the tooltip I discovered
that there was a race-condition in the ChartGroup tests which meant
that only the first chart test was working correctly, which explained
why it was significantly slower than the others. This was due to
how the animation completed marker was set, which is required due
to how Recharts works.

Since the marker was only ever set, and never removed, and since Euclid
is a SPA, and no reloading was done, the marker was set correctly on the
first test. Then subsequent tests simply checked for it and immediately
found that it was set (incorrectly) and exited. Hence, the
race-condition. This marker is now unset when data is not finished loading
from Apollo and this fixes the problem.

Signed-off-by: David Piper <david.piper@mailfence.com>