fix(client): resolve problem where the chart kept reloading forever
The Storybook visual regression tests (ironically) introduced
a regression... Specifically, the marker used to indicate that the
chart has finished animating, so that a snapshot can be taken,
made the chart reload forever. This was due to how React detects
whether or not something should re-render. The setting of the
marker (even if it was set to the same value) forced a re-render.

To solve this, I simply only set it if it would be set to a different
value.

Signed-off-by: David Piper <david.piper@mailfence.com>