refactor(client): stop using apollo-boost, to prepare for speed-up
In an effort to speed up the GraphQL queries, I am preparing
to start using Apollo Link Batch, to batch the queries to
avoid extraneous queries. To do so, I have to first eject
from apollo-boost which is what this does.

The chart animation time is also being lowered from 1.5 seconds
to 0.5 seconds, to help make it more responsive.

Signed-off-by: David Piper <david.piper@mailfence.com>