fix(client): fix the client from reloading the chart (again)
The last fix of the chart's infinite reloading, had a bug
in it, which as you might have guessed; undid the other
fix... This resolves that other bug, which was I was unsetting
the variable indicating that the loading was done, so it
just kept reloading.

I have noticed that things are a bit faster now, so
this was likely affecting the performance, despite the
fact that the performance being measured in the e2e
test is just the ticker load time, since they are
coming from the same server this _could_ have potentially
affected it somewhat.

Signed-off-by: David Piper <david.piper@mailfence.com>