test: reduce cypress timeout back to default
I had increased the cypress timeout dramatically due to the
incredibly slow loading of Euclid issue. Now that this has
been fixed, I am lowering it back to the defaults. This will
help to ensure that the tests are ensuring that Euclid is
fast, and usable as well as technically functioning (but
potentially incredibly slow; as was the case until now).

Signed-off-by: David Piper <david.piper@mailfence.com>