build(root): expose cypress npm task; found during code review
Sometimes, it is useful to be able to just run the Cypress integration
and e2e tests by themselves. So I have exposed this as an npm script.
I had accidentally left this in as an exposed gulp task during
development and I decided to go ahead and just make it official
while doing a code review.

Signed-off-by: David Piper <david.piper@mailfence.com>