style(server): add TS definitions for the crawler module
The crawler module is written in vanilla JS, so I have written
my own. This should be moved into DefinitelyTyped so that
everyone in the community can benefit from these definitions
and not just me, as I don't want to be selfish and a code-hog.

To facilitate this an issue should be made in the DefinitelyTyped
Epic (https://github.com/dbpiper/Euclid/issues/64).

Signed-off-by: David Piper <david.piper@mailfence.com>