feat(server): make amazon product crawler
Complete work on an initial version of the Amazon product
crawler, it wraps the `crawl` module and simply extracts the
desired product information from the queued pages. Pages
can continue to be queued indefinitely, which means that
theoretically it could simply run on its own server as a service
and continually collect data.

I am not sure if this is what we want, however I do know
that it will take a long time to collect the data, and crawling
it is not really the purpose of Euclid. Instead the purpose of
Euclid is displaying the data, so it might be a good idea
to simply take this part and make it a separate project. At
the very least, I plan on putting this Amazon crawler into its
own module.

Signed-off-by: David Piper <david.piper@mailfence.com>