refactor(crawler): have scraper return promise, make crawler separate
The scraper wasn't returning anything previously, now it returns
a Promise that is resolved with either the AmazonProduct information
or an Error. It doesn't have to return an Observable since it is only
scraping one product. The crawler will handle multiple products
and this also makes that exist in a more formal way; in its own file
rather than just sitting in index.ts.

Signed-off-by: David Piper <david.piper@mailfence.com>